rename listener and add separate listener for release
